PF-LOGIN: move route in const file. Use privacy-policy route from pn-common
