+++ packages/pn-personafisica-login/src/navigation/routes.const.ts
@@ -1,0 +1,4 @@
+export const ROUTE_LOGOUT = '/logout';
+export const ROUTE_LOGIN = '/login';
+export const ROUTE_LOGIN_ERROR = '/login/error';
+export const ROUTE_SUCCESS = '/login/success';
