+++ packages/pn-personafisica-login/src/pages/login/Login.tsx
@@ -7,7 +7,12 @@ import Button from '@mui/material/Button';
 import Grid from '@mui/material/Grid';
 import Typography from '@mui/material/Typography';
 import { styled } from '@mui/material/styles';
-import { AppRouteParams, Layout, useIsMobile } from '@pagopa-pn/pn-commons';
+import {
+  AppRouteParams,
+  Layout,
+  PRIVACY_LINK_RELATIVE_PATH as PRIVACY_POLICY,
+  useIsMobile,
+} from '@pagopa-pn/pn-commons';
 import { CieIcon, SpidIcon } from '@pagopa/mui-italia/dist/icons';
 
 import { PFLoginEventsType } from '../../models/PFLoginEventsType';
@@ -31,7 +36,7 @@ const Login = () => {
   const [params] = useSearchParams();
   const aar = params.get(AppRouteParams.AAR);
   const { URL_API_LOGIN, SPID_CIE_ENTITY_ID, PAGOPA_HELP_EMAIL, PF_URL } = getConfiguration();
-  const ROUTE_PRIVACY_POLICY = PF_URL + '/informativa-privacy';
+  const privacyPolicyUrl = `${PF_URL}${PRIVACY_POLICY}`;
 
   if (aar !== null && aar !== '') {
     storageAarOps.write(aar);
@@ -80,7 +85,7 @@ const Login = () => {
         surname: undefined,
         email: undefined,
       }}
-      privacyPolicyHref={ROUTE_PRIVACY_POLICY}
+      privacyPolicyHref={privacyPolicyUrl}
     >
       <Grid container direction="column" my={isMobile ? 4 : 16} alignItems="center" id="loginPage">
         <Grid item>
